+++ src/onegov.ballot/onegov/ballot/__init__.py
@@ -1,4 +1,10 @@
 from onegov.ballot.models import Ballot, BallotResult, Vote
-from onegov.ballot.collection import VoteCollection
+from onegov.ballot.collection import BallotCollection, VoteCollection
 
-__all__ = ['Ballot', 'BallotResult', 'Vote', 'VoteCollection']
+__all__ = [
+    'Ballot',
+    'BallotCollection',
+    'BallotResult',
+    'Vote',
+    'VoteCollection'
+]