+++ src/onegov.ballot/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+- Each ballot result now needs a municipality id, a.k.a BFS-Nummer.
+  [href]
+
 0.0.1 (2015-06-18)
 ~~~~~~~~~~~~~~~~~~~